refactor(logging): split persistence delivery ownership
